+++ packages/presets/src/theme/recipes/switch.ts
@@ -20,10 +20,10 @@ export const switchRecipe = defineRecipe({
       display: 'inline-flex',
       flexShrink: '0',
       transitionDuration: 'normal',
-      transitionProperty: 'background-color',
+      transitionProperty: 'background',
       transitionTimingFunction: 'default',
       _checked: {
-        background: 'bg.accent',
+        background: 'accent.default',
       },
     },
     label: {
@@ -36,7 +36,7 @@ export const switchRecipe = defineRecipe({
       borderRadius: 'full',
       boxShadow: 'xs',
       transitionDuration: 'normal',
-      transitionProperty: 'transform, background-color',
+      transitionProperty: 'transform, background',
       transitionTimingFunction: 'default',
       zIndex: '1',
       _checked: {