+++ packages/presets/src/theme/semantic-tokens.ts
@@ -2,31 +2,42 @@ import { defineSemanticTokens } from '@pandacss/dev'
 
 export const semanticTokens = defineSemanticTokens({
   colors: {
-    fg: {
-      default: { value: { base: '{colors.neutral.950}', _dark: '{colors.white}' } },
-      emphasized: { value: { base: '{colors.neutral.700}', _dark: '{colors.neutral.200}' } },
-      muted: { value: { base: '{colors.neutral.600}', _dark: '{colors.neutral.300}' } },
-      subtle: { value: { base: '{colors.neutral.500}', _dark: '{colors.neutral.400}' } },
-      accent: { value: { base: '{colors.white}', _dark: '{colors.neutral.950}' } },
-      disabled: { value: { base: '{colors.neutral.200}', _dark: '{colors.neutral.800}' } },
-    },
     bg: {
       canvas: { value: { base: '{colors.neutral.25}', _dark: '{colors.neutral.950}' } },
       default: { value: { base: '{colors.white}', _dark: '{colors.neutral.950}' } },
       subtle: { value: { base: '{colors.neutral.100}', _dark: '{colors.neutral.900}' } },
       muted: { value: { base: '{colors.neutral.200}', _dark: '{colors.neutral.800}' } },
       emphasized: { value: { base: '{colors.neutral.300}', _dark: '{colors.neutral.700}' } },
-      accent: { value: { base: '{colors.neutral.950}', _dark: '{colors.white}' } },
       disabled: { value: { base: '{colors.neutral.200}', _dark: '{colors.neutral.800}' } },
     },
+    fg: {
+      default: { value: { base: '{colors.neutral.950}', _dark: '{colors.white}' } },
+      emphasized: { value: { base: '{colors.neutral.700}', _dark: '{colors.neutral.200}' } },
+      muted: { value: { base: '{colors.neutral.600}', _dark: '{colors.neutral.300}' } },
+      subtle: { value: { base: '{colors.neutral.500}', _dark: '{colors.neutral.400}' } },
+      disabled: { value: { base: '{colors.neutral.200}', _dark: '{colors.neutral.800}' } },
+    },
+    accent: {
+      default: { value: { base: '{colors.neutral.950}', _dark: '{colors.white}' } },
+      emphasized: { value: { base: '{colors.neutral.800}', _dark: '{colors.neutral.200}' } },
+      fg: { value: { base: '{colors.white}', _dark: '{colors.neutral.950}' } },
+    },
+
     border: {
       default: { value: { base: '{colors.neutral.200}', _dark: '{colors.neutral.800}' } },
       emphasized: { value: { base: '{colors.neutral.300}', _dark: '{colors.neutral.700}' } },
+      outline: { value: { base: '{colors.neutral.700}', _dark: '{colors.neutral.300}' } },
       accent: { value: { base: '{colors.neutral.950}', _dark: '{colors.white}' } },
       disabled: { value: { base: '{colors.neutral.200}', _dark: '{colors.neutral.800}' } },
     },
   },
   shadows: {
+    outline: {
+      value: {
+        base: '0 0 0 1px {colors.border.outline}',
+        _dark: '0 0 0 1px {colors.border.outline}',
+      },
+    },
     xs: {
       value: {
         base: '0px 1px 2px rgba(23, 23, 23,  0.1)',