+++ website/src/components/docs/components/demo/number-input-demo.tsx
@@ -4,8 +4,8 @@ import {
   NumberInput,
   NumberInputControl,
   NumberInputDecrementTrigger,
-  NumberInputField,
   NumberInputIncrementTrigger,
+  NumberInputInput,
   NumberInputScrubber,
 } from '~/components/number-input'
 
@@ -13,7 +13,7 @@ export const NumberInputDemo = () => {
   return (
     <NumberInput min={-50} max={50} defaultValue="42">
       <NumberInputScrubber />
-      <NumberInputField />
+      <NumberInputInput />
       <NumberInputControl>
         <NumberInputIncrementTrigger asChild>
           <IconButton size="xs" variant="tertiary" icon={<FiChevronUp />} aria-label="Increment" />