+++ website/src/components/docs/components/demo/pin-input-demo.tsx
@@ -2,7 +2,7 @@ import { Input } from '~/components/input'
 import {
   PinInput,
   PinInputControl,
-  PinInputField,
+  PinInputInput,
   type PinInputProps,
 } from '~/components/pin-input'
 
@@ -10,9 +10,9 @@ export const PinInputDemo = (props: PinInputProps) => (
   <PinInput placeholder="0" onComplete={(e) => alert(e.valueAsString)}>
     <PinInputControl>
       {[0, 1, 2, 3].map((id, index) => (
-        <PinInputField key={id} index={index} asChild>
+        <PinInputInput key={id} index={index} asChild>
           <Input size="xl" width="0" textAlign="center" />
-        </PinInputField>
+        </PinInputInput>
       ))}
     </PinInputControl>
   </PinInput>