+++ website/src/components/marketing/examples/marketing/page-header.tsx
@@ -5,7 +5,11 @@ export const PageHeader = () => {
   return (
     <Stack gap={{ base: '4', md: '6' }} align="center" textAlign="center">
       <Stack gap="3" align="center">
-        <Typography textStyle={{ base: 'sm', md: 'md' }} fontWeight="semibold" color="accent">
+        <Typography
+          textStyle={{ base: 'sm', md: 'md' }}
+          fontWeight="semibold"
+          color="accent.default"
+        >
           Pricing
         </Typography>
         <Heading textStyle={{ base: '4xl', md: '5xl' }}>Choose Your Path to Success</Heading>